requirements.txt
lightweight-charts
backtrader 
Backtesting 
ccxt 
requests 
pandas 
yfinance 
scikit-learn 
numpy 
matplotlib
pandas-ta
termcolor
websockets
yahooquery
psycopg2-binary
ib_insync
ibapi
protobuf
futu-api
ta
TA-Lib